LPD-42633 This method is generic
